GameDemo.cpp: 修正include路径；修改参数double为float ChsHUDManager.cpp：修正include路径 ChsRenderSystem.cpp：去掉类型转换的warning ChsSprite2D.cpp：去掉类型转换的warning；去掉匿名数组 ChsSprite2D.h：文件编码修改为UTF-8 ChsTexture2D.cpp：currentTextureHandle为unsigned类型不能为-1 ChsQuaternion.cpp：文件编码修改为UTF-8 ChsQuaternion.h：文件编码修改为UTF-8 ChsVector3.cpp：文件编码修改为UTF-8 ChsVector3.h：文件编码修改为UTF-8 ChsShaderProgram.h：去参数未使用的warning ChsShaderUniform.h：去掉无用的函数

+++ src/chaos/ChsRenderSystem.cpp
@@ -225,10 +225,10 @@ namespace Chaos {
 	
   //------------------------------------------------------------------------------------------------
 	void ChsRenderSystem::setViewPort( int x, int y, int w, int h ){
-		this->viewport.x = x;
-		this->viewport.y = y;
-		this->viewport.w = w;
-		this->viewport.h = h;
+		this->viewport.x = static_cast<float>( x );
+		this->viewport.y = static_cast<float>( y );
+		this->viewport.w = static_cast<float>( w );
+		this->viewport.h = static_cast<float>( h );
 		glViewport( x, y, w, h );
 	}
